+++ cloup/constraints/_support.py
@@ -201,16 +201,13 @@ class ConstraintMixin:
             with formatter.section('Constraints'):
                 formatter.write_dl(records)
 
-    def format_options(self, ctx, formatter: HelpFormatter) -> None:
-        super().format_options(ctx, formatter)  # type: ignore
+    def must_show_constraints(self, ctx: Context) -> bool:
         # By default, don't show constraints
-        must_show_constraints = first_bool(
+        return first_bool(
             self.show_constraints,
             getattr(ctx, "show_constraints", None),
             False,
         )
-        if must_show_constraints:
-            self.format_constraints(ctx, formatter)
 
 
 def ensure_constraints_support(command) -> ConstraintMixin: